refactor: export content package
